chore: Merged DocumentCreationService
