--- src/test/kotlin/nl/info/zac/documentcreation/CmmnDocumentCreationServiceTest.kt
@@ -1,198 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2025 INFO.nl
- * SPDX-License-Identifier: EUPL-1.2+
- */
-
-package nl.info.zac.documentcreation
-
-import io.kotest.core.spec.style.BehaviorSpec
-import io.kotest.matchers.shouldBe
-import io.mockk.checkUnnecessaryStub
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.slot
-import jakarta.enterprise.inject.Instance
-import nl.info.client.smartdocuments.model.document.Data
-import nl.info.client.smartdocuments.model.document.SmartDocument
-import nl.info.client.zgw.model.createZaak
-import nl.info.zac.app.informatieobjecten.EnkelvoudigInformatieObjectUpdateService
-import nl.info.zac.authentication.LoggedInUser
-import nl.info.zac.authentication.createLoggedInUser
-import nl.info.zac.configuration.ConfigurationService
-import nl.info.zac.documentcreation.converter.DocumentCreationDataConverter
-import nl.info.zac.documentcreation.model.createCmmnDocumentCreationDataAttended
-import nl.info.zac.documentcreation.model.createData
-import nl.info.zac.documentcreation.model.createDocumentCreationAttendedResponse
-import nl.info.zac.smartdocuments.SmartDocumentsService
-import nl.info.zac.smartdocuments.SmartDocumentsTemplatesService
-import java.net.URI
-import java.net.URLEncoder
-import java.time.ZoneOffset
-import java.time.ZonedDateTime
-import java.time.format.DateTimeFormatter
-import java.util.UUID
-
-class CmmnDocumentCreationServiceTest : BehaviorSpec({
-    val smartDocumentsService = mockk<SmartDocumentsService>()
-    val smartDocumentsTemplatesService = mockk<SmartDocumentsTemplatesService>()
-    val documentCreationDataConverter = mockk<DocumentCreationDataConverter>()
-    val loggedInUserInstance = mockk<Instance<LoggedInUser>>()
-    val enkelvoudigInformatieObjectUpdateService = mockk<EnkelvoudigInformatieObjectUpdateService>()
-    val configurationService: ConfigurationService = mockk<ConfigurationService>()
-    val documentCreationService = DocumentCreationService(
-        smartDocumentsService = smartDocumentsService,
-        documentCreationDataConverter = documentCreationDataConverter,
-        enkelvoudigInformatieObjectUpdateService = enkelvoudigInformatieObjectUpdateService,
-        configurationService = configurationService
-    )
-    val cmmnDocumentCreationService = CmmnDocumentCreationService(
-        smartDocumentsService = smartDocumentsService,
-        smartDocumentsTemplatesService = smartDocumentsTemplatesService,
-        documentCreationDataConverter = documentCreationDataConverter,
-        documentCreationService = documentCreationService,
-        loggedInUserInstance = loggedInUserInstance
-    )
-
-    beforeEach {
-        checkUnnecessaryStub()
-    }
-
-    Given("Document creation data with a zaak and an information object type") {
-        val zaakTypeUUID = UUID.randomUUID()
-        val zaakTypeURI = URI("https://example.com/$zaakTypeUUID")
-        val zaak = createZaak(zaaktypeUri = zaakTypeURI)
-        val taskId = "fakeTaskId"
-        val documentCreationData = createCmmnDocumentCreationDataAttended(
-            zaak = zaak,
-            taskId = taskId
-        )
-        val userDisplayName = "fakeDisplayName"
-        val loggedInUser = createLoggedInUser(
-            displayName = userDisplayName
-        )
-        val data = createData()
-        val documentCreationAttendedResponse = createDocumentCreationAttendedResponse()
-        val contextUrl = "https://example.com"
-        val templateGroupName = "fakeTemplateGroupName"
-        val templateName = "fakeTemplateName"
-        val dataSlot = slot<Data>()
-        val smartDocumentSlot = slot<SmartDocument>()
-
-        every { loggedInUserInstance.get() } returns loggedInUser
-        every {
-            documentCreationDataConverter.createData(
-                loggedInUser,
-                documentCreationData.zaak,
-                documentCreationData.taskId
-            )
-        } returns data
-        every {
-            smartDocumentsService.createDocumentAttended(capture(dataSlot), capture(smartDocumentSlot))
-        } returns documentCreationAttendedResponse
-        every {
-            smartDocumentsTemplatesService.getTemplateGroupName(documentCreationData.templateGroupId)
-        } returns templateGroupName
-        every {
-            smartDocumentsTemplatesService.getTemplateName(documentCreationData.templateId)
-        } returns templateName
-        every { configurationService.readContextUrl() } returns contextUrl
-
-        When("the 'create document attended' method is called") {
-            val documentCreationResponse = cmmnDocumentCreationService.createCmmnDocumentAttended(documentCreationData)
-
-            Then(
-                """
-                the smart documents service is called to create an attended document and a document creation response is returned
-                """
-            ) {
-                with(documentCreationResponse) {
-                    redirectUrl shouldBe documentCreationAttendedResponse.redirectUrl
-                    message shouldBe documentCreationAttendedResponse.message
-                }
-                with(smartDocumentSlot.captured) {
-                    selection.templateGroup shouldBe templateGroupName
-                    selection.template shouldBe templateName
-                    with(variables!!) {
-                        outputFormats.size shouldBe 1
-                        outputFormats[0].outputFormat shouldBe "docx"
-                        redirectMethod shouldBe "POST"
-                        redirectUrl shouldBe "$contextUrl/rest/document-creation/smartdocuments/callback" +
-                            "/zaak/${zaak.uuid}" +
-                            "/task/$taskId" +
-                            "?title=${URLEncoder.encode(documentCreationData.title, Charsets.UTF_8)}" +
-                            "&userName=${URLEncoder.encode(userDisplayName, Charsets.UTF_8)}" +
-                            "&creationDate=${URLEncoder.encode(
-                                documentCreationData.creationDate.format(DateTimeFormatter.ISO_OFFSET_DATE_TIME),
-                                Charsets.UTF_8
-                            )}" +
-                            "&templateId=" +
-                            "${URLEncoder.encode(documentCreationData.templateId, Charsets.UTF_8)}" +
-                            "&templateGroupId=" +
-                            "${URLEncoder.encode(documentCreationData.templateGroupId, Charsets.UTF_8)}"
-                    }
-                }
-            }
-        }
-    }
-
-    Given("A zaak exists") {
-        val contextUrl = "https://example.com:2222"
-        val zaakUuid = UUID.randomUUID()
-        val templateGroupId = "groupId"
-        val templateId = "templateId"
-        val title = "title"
-        val description = "description"
-        val creationDate = ZonedDateTime.of(2024, 10, 7, 0, 0, 0, 0, ZoneOffset.UTC)
-        val userName = "Full User Name"
-
-        every { configurationService.readContextUrl() } returns contextUrl
-
-        When("Document creation URL is requested for zaak") {
-            val uri = cmmnDocumentCreationService.documentCreationCallbackUrl(
-                zaakUuid = zaakUuid,
-                null,
-                templateGroupId,
-                templateId,
-                title,
-                description,
-                creationDate,
-                userName
-            )
-
-            Then("Correct URl is provided") {
-                uri.toString() shouldBe "$contextUrl/rest/document-creation/smartdocuments/callback/zaak/$zaakUuid" +
-                    "?title=$title" +
-                    "&userName=Full+User+Name" +
-                    "&creationDate=2024-10-07T00%3A00%3A00Z" +
-                    "&description=$description" +
-                    "&templateId=$templateId" +
-                    "&templateGroupId=$templateGroupId"
-            }
-        }
-
-        When("Document creation URL is requested for taak") {
-            val taakUuid = UUID.randomUUID().toString()
-            val uri = cmmnDocumentCreationService.documentCreationCallbackUrl(
-                zaakUuid,
-                taakUuid,
-                templateGroupId,
-                templateId,
-                title,
-                description,
-                creationDate,
-                userName
-            )
-
-            Then("Correct URl is provided") {
-                uri.toString() shouldBe
-                    "$contextUrl/rest/document-creation/smartdocuments/callback/zaak/$zaakUuid/task/$taakUuid" +
-                    "?title=$title" +
-                    "&userName=Full+User+Name" +
-                    "&creationDate=2024-10-07T00%3A00%3A00Z" +
-                    "&description=$description" +
-                    "&templateId=$templateId" +
-                    "&templateGroupId=$templateGroupId"
-            }
-        }
-    }
-})
